añadio primer frontend

+++ src/main/java/com/example/demo/repository/CamaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Cama;
 
-public interface CamaRepository extends CrudRepository<Cama,Long>{
+public interface CamaRepository extends CrudRepository<Cama,Integer>{
 
 }
