+++ src/main/java/com/example/demo/repository/VacunaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Vacuna;
 
-public interface VacunaRepository extends CrudRepository<Vacuna,Long>{
+public interface VacunaRepository extends CrudRepository<Vacuna,Integer>{
 
 }
